+++ matchers.go
@@ -37,50 +37,3 @@ func (m RequestMatcher) WithJSONFields(json map[string]interface{}) RequestMatch
 	}
 	return m
 }
-
-type Expectation struct {
-	Request  RequestMatcher      `json:"httpRequest,omitempty"`
-	Response ExpectationResponse `json:"httpResponse,omitempty"`
-	Priority int32               `json:"priority,omitempty"`
-}
-
-func NewExpectation(request RequestMatcher) *Expectation {
-	e := new(Expectation)
-	e.Request = request
-	return e
-}
-
-func (e Expectation) WithResponse(response ExpectationResponse) Expectation {
-	e.Response = response
-	return e
-}
-
-func (e Expectation) WithPriority(priority int32) Expectation {
-	e.Priority = priority
-	return e
-}
-
-type ExpectationResponse struct {
-	Body       map[string]interface{} `json:"body,omitempty"`
-	Headers    map[string][]string    `json:"headers,omitempty"`
-	StatusCode int32                  `json:"statusCode,omitempty"`
-}
-
-func NewResponseOK() *ExpectationResponse {
-	e := new(ExpectationResponse)
-	e.StatusCode = 200
-	return e
-}
-
-func (e ExpectationResponse) WithJSONBody(json map[string]interface{}) ExpectationResponse {
-	e.Body = json
-	return e
-}
-
-func (e ExpectationResponse) WithHeader(key, value string) ExpectationResponse {
-	if e.Headers == nil {
-		e.Headers = make(map[string][]string)
-	}
-	e.Headers[key] = []string{value}
-	return e
-}
